+++ app/models/null_table.rb
@@ -1,0 +1,17 @@
+class NullTable
+  def columns
+    []
+  end
+
+  def column_ids
+    []
+  end
+
+  def name
+    '-'
+  end
+
+  def calculations
+    Calculation.none
+  end
+end