+++ test/unit/spreadsheet_test.rb
@@ -29,80 +29,32 @@ module TableCaclulation
              :members, :member_roles, :roles, :users,
              :tables, :spreadsheets
 
+    test 'should respond to safe attributes' do
+      assert Spreadsheet.respond_to? :safe_attributes
+    end
 
-    # test 'should have many columns' do
-    #   association = Table.reflect_on_association(:columns)
-    #   assert_equal :columns, association.name
-    #   assert_equal :has_and_belongs_to_many, association.macro
-    # end
+    test 'should respond to column_ids' do
+      assert Spreadsheet.find(1).respond_to? :column_ids
+    end
 
-    # test 'should have many project types' do
-    #   association = Table.reflect_on_association(:project_types)
-    #   assert_equal :project_types, association.name
-    #   assert_equal :has_and_belongs_to_many, association.macro
-    # end
+    test 'should respond to calculations' do
+      assert Spreadsheet.find(1).respond_to? :calculations?
+    end
 
-    # test 'should respond to safe attributes' do
-    #   assert Table.respond_to? :safe_attributes
-    # end
+    test 'should respond to copy' do
+      assert Spreadsheet.find(1).respond_to? :copy
+    end
 
-    # test 'should respond to column_assigned?' do
-    #   assert Table.find(1).respond_to? :column_assigned?
-    # end
+    test 'spreadsheet should not be valid without name' do
+      sheet = Spreadsheet.new(table_id: Table.find(1).id)
+      assert_not sheet.valid?
+      assert_equal %i[name], sheet.errors.keys
+    end
 
-    # test 'should respond to project_type_assigned?' do
-    #   assert Table.find(1).respond_to? :project_type_assigned?
-    # end
-
-    # test 'table should not be valid without name' do
-    #   table = Table.new
-    #   assert_not table.valid?
-    #   assert_equal [:name], table.errors.keys
-    # end
-
-    # test 'should find columns and project types' do
-    #   cf = custom_field
-    #   table = Table.find(1)
-    #   table.columns << cf
-    #   assert table.save
-    #   assert table.columns.count == 1
-    #   pt = find_project_type(id: 4)
-    #   table.project_types << pt
-    #   assert table.save
-    #   assert table.project_types.count == 1
-    # end
-
-    # test 'should update table' do
-    #   table = Table.new(name: 'Table to update')
-    #   name = 'Updated Table'
-    #   table.name = name
-    #   assert table.save
-    #   assert table.name == name
-    # end
-
-    # test 'should delete table' do
-    #   table = Table.new(name: 'Table to delete')
-    #   assert table.destroy
-    #   assert table.destroyed?
-    # end
-
-    # private
-
-    # def custom_field
-    #   CustomField.generate! custom_field_attributes(name: 'Table Column')
-    # end
-
-    # def custom_field_attributes(name:)
-    #   { name: name,
-    #     regexp: '',
-    #     is_for_all: true,
-    #     is_filter: true,
-    #     type: 'TableCustomField',
-    #     possible_values: %w[A B C],
-    #     is_required: false,
-    #     field_format: 'list',
-    #     default_value: '',
-    #     editable: true }
-    # end
+    test 'spreadsheet should not be valid without table' do
+      sheet = Spreadsheet.new(name: 'New Sheet')
+      assert_not sheet.valid?
+      assert_equal %i[table_id], sheet.errors.keys
+    end
   end
 end